# openapi.json
{
  "components" : {
    "schemas" : {
...
      "PageResponsePlatformTicketRow" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/PlatformTicketRow"
            },
            "type" : [ "array", "null" ]
          },
          "last" : {
            "type" : "boolean"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "last", "page", "size", "total" ],
        "type" : "object"
      },
...
      "PlatformBusinessDetail" : {
        "properties" : {
          "country" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "currency" : {
            "type" : [ "string", "null" ]
          },
          "currentPeriodEnd" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "devices" : {
            "format" : "int64",
            "type" : "integer"
          },
          "effectivePlan" : {
            "type" : [ "string", "null" ]
          },
          "flags" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : [ "object", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "members" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "openTickets" : {
            "format" : "int64",
            "type" : "integer"
          },
          "owners" : {
            "items" : {
              "$ref" : "#/components/schemas/PlatformOwner"
            },
            "type" : [ "array", "null" ]
          },
          "plan" : {
            "type" : [ "string", "null" ]
          },
          "planNote" : {
            "type" : [ "string", "null" ]
          },
          "planStatus" : {
            "type" : [ "string", "null" ]
          },
          "sales30" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : [ "string", "null" ]
          },
          "suspendedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "suspendedReason" : {
            "type" : [ "string", "null" ]
          },
          "timezone" : {
            "type" : [ "string", "null" ]
          },
          "trialEndsAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "type" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "devices", "id", "members", "openTickets", "sales30" ],
        "type" : "object"
      },
...
      "PlatformOwner" : {
        "properties" : {
          "email" : {
            "type" : [ "string", "null" ]
          },
          "fullName" : {
            "type" : [ "string", "null" ]
          },
          "userId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
      "PlatformTicketRow" : {
        "properties" : {
          "businessId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "businessName" : {
            "type" : [ "string", "null" ]
          },
          "category" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "diagnostics" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "locale" : {
            "type" : [ "string", "null" ]
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "replyToEmail" : {
            "type" : [ "string", "null" ]
          },
          "replyToPhone" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
...
      "TicketCreated" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reference" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
      "TicketRequest" : {
        "properties" : {
          "businessId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "category" : {
            "pattern" : "QUESTION|PROBLEM|SUGGESTION|BILLING",
            "type" : [ "string", "null" ]
          },
          "diagnostics" : {
            "maxLength" : 20000,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "locale" : {
            "pattern" : "es|en",
            "type" : [ "string", "null" ]
          },
          "message" : {
            "maxLength" : 4000,
            "minLength" : 10,
            "type" : [ "string", "null" ]
          },
          "replyToEmail" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "replyToPhone" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "message" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/platform/businesses/{id}" : {
      "get" : {
        "operationId" : "business",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlatformBusinessDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "platform-controller" ]
      }
    },
...
    "/api/platform/tickets" : {
      "get" : {
        "operationId" : "tickets",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 25,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePlatformTicketRow"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "platform-controller" ]
      }
    },
...
    "/api/support/tickets" : {
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TicketRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketCreated"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "support-controller" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}